pip-requirements.txt
annotated-types==0.6.0
blinker==1.7.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.3
cymem==2.0.8
de-core-news-md @ https://github.com/explosion/spacy-models/releases/download/de_core_news_md-3.7.0/de_core_news_md-3.7.0-py3-none-any.whl#sha256=1426896f135b7e6314637faa9025a3e580c9ba2785372ccffe723dc20b41c493
de-core-news-sm @ https://github.com/explosion/spacy-models/releases/download/de_core_news_sm-3.7.0/de_core_news_sm-3.7.0-py3-none-any.whl#sha256=d88c737eb7eb766f730f6a2dcb99dfcdb81623e1e0d89a9c638a2182ac19c52e
flask==3.0.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
langcodes==3.3.0
MarkupSafe==2.1.3
murmurhash==1.0.10
numpy==1.26.1
opencv-python==4.8.1.78
packaging==23.2
pdf2image==1.16.3
Pillow==10.1.0
preshed==3.0.9
pycryptodome==3.21.0
pydantic==2.4.2
pydantic_core==2.10.1
pytesseract==0.3.10
requests==2.31.0
setuptools==68.2.2
smart-open==6.4.0
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
tesseract==0.1.3
pytesseract
thinc==8.2.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
waitress==3.0.2
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.1
